Tools: Moved boost and series modules.
