+++ js/Extensions/Boost/BoostInit.js
@@ -15,8 +15,8 @@ import H from '../../Core/Globals.js';
 import U from '../../Core/Utilities.js';
 var addEvent = U.addEvent, extend = U.extend, fireEvent = U.fireEvent, wrap = U.wrap;
 import '../../Core/Series/Series.js';
-import butils from './boost-utils.js';
-import createAndAttachRenderer from './boost-attach.js';
+import butils from './BoostUtils.js';
+import createAndAttachRenderer from './BoostAttach.js';
 var Series = H.Series, seriesTypes = H.seriesTypes, noop = function () { }, eachAsync = butils.eachAsync, pointDrawHandler = butils.pointDrawHandler, allocateIfNotSeriesBoosting = butils.allocateIfNotSeriesBoosting, renderIfNotSeriesBoosting = butils.renderIfNotSeriesBoosting, shouldForceChartSeriesBoosting = butils.shouldForceChartSeriesBoosting, index;
 /* eslint-disable valid-jsdoc */
 /**
