+++ js/Extensions/Boost/BoostOptions.js
