+++ js/Extensions/Boost/BoostOverrides.js
@@ -9,74 +9,19 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-
 'use strict';
-
-import type SVGPath from '../../Core/Renderer/SVG/SVGPath';
 import Chart from '../../Core/Chart/Chart.js';
 import H from '../../Core/Globals.js';
 import Point from '../../Core/Series/Point.js';
 import U from '../../Core/Utilities.js';
-const {
-    addEvent,
-    error,
-    getOptions,
-    isArray,
-    isNumber,
-    pick,
-    wrap
-} = U;
-
-/**
- * Internal types
- * @private
- */
-declare global {
-    namespace Highcharts {
-        interface BoostAlteredObject {
-            own: boolean;
-            prop: string;
-            val: unknown;
-            value: unknown;
-        }
-        interface ChartLike {
-            /** @requires modules/boost */
-            getBoostClipRect(target: BoostTargetObject): BBoxObject;
-            /** @requires modules/boost */
-            isChartSeriesBoosting(): boolean;
-        }
-        interface Series {
-            alteredByBoost?: Array<BoostAlteredObject>;
-            /** @requires modules/boost */
-            destroyGraphics(): void;
-            /** @requires modules/boost */
-            enterBoost(): void;
-            /** @requires modules/boost */
-            exitBoost(): void;
-            /** @requires modules/boost */
-            hasExtremes(checkX?: boolean): boolean;
-        }
-        interface SeriesOptions {
-            boostData?: Array<unknown>;
-        }
-    }
-}
-
+var addEvent = U.addEvent, error = U.error, getOptions = U.getOptions, isArray = U.isArray, isNumber = U.isNumber, pick = U.pick, wrap = U.wrap;
 import '../../Core/Series/Series.js';
 import '../../Core/Options.js';
-
 import '../../Core/Interaction.js';
-
-import butils from './boost-utils.js';
-import boostable from './boostables.js';
-import boostableMap from './boostable-map.js';
-
-var boostEnabled = butils.boostEnabled,
-    shouldForceChartSeriesBoosting = butils.shouldForceChartSeriesBoosting,
-    Series = H.Series,
-    seriesTypes = H.seriesTypes,
-    plotOptions: Highcharts.PlotOptions = getOptions().plotOptions as any;
-
+import butils from './BoostUtils.js';
+import boostable from './Boostables.js';
+import boostableMap from './BoostableMap.js';
+var boostEnabled = butils.boostEnabled, shouldForceChartSeriesBoosting = butils.shouldForceChartSeriesBoosting, Series = H.Series, seriesTypes = H.seriesTypes, plotOptions = getOptions().plotOptions;
 /**
  * Returns true if the chart is in series boost mode.
  *
@@ -88,21 +33,13 @@ var boostEnabled = butils.boostEnabled,
  * @return {boolean}
  *         true if the chart is in series boost mode
  */
-Chart.prototype.isChartSeriesBoosting = function (): boolean {
-    var isSeriesBoosting: boolean,
-        threshold = pick(
-            this.options.boost && this.options.boost.seriesThreshold,
-            50
-        );
-
+Chart.prototype.isChartSeriesBoosting = function () {
+    var isSeriesBoosting, threshold = pick(this.options.boost && this.options.boost.seriesThreshold, 50);
     isSeriesBoosting = threshold <= this.series.length ||
         shouldForceChartSeriesBoosting(this);
-
     return isSeriesBoosting;
 };
-
 /* eslint-disable valid-jsdoc */
-
 /**
  * Get the clip rectangle for a target, either a series or the chart. For the
  * chart, we need to consider the maximum extent of its Y axes, in case of
@@ -115,29 +52,21 @@ Chart.prototype.isChartSeriesBoosting = function (): boolean {
  *
  * @return {Highcharts.BBoxObject}
  */
-Chart.prototype.getBoostClipRect = function (
-    target: Chart
-): Highcharts.BBoxObject {
+Chart.prototype.getBoostClipRect = function (target) {
     var clipBox = {
         x: this.plotLeft,
         y: this.plotTop,
         width: this.plotWidth,
         height: this.plotHeight
     };
-
     if (target === this) {
-        this.yAxis.forEach(function (yAxis: Highcharts.Axis): void {
+        this.yAxis.forEach(function (yAxis) {
             clipBox.y = Math.min(yAxis.pos, clipBox.y);
-            clipBox.height = Math.max(
-                yAxis.pos - this.plotTop + yAxis.len,
-                clipBox.height
-            );
+            clipBox.height = Math.max(yAxis.pos - this.plotTop + yAxis.len, clipBox.height);
         }, this);
     }
-
     return clipBox;
 };
-
 /**
  * Return a full Point object based on the index.
  * The boost module uses stripped point objects for performance reasons.
@@ -150,29 +79,16 @@ Chart.prototype.getBoostClipRect = function (
  * @return {Highcharts.Point}
  *         A Point object as per https://api.highcharts.com/highcharts#Point
  */
-Series.prototype.getPoint = function (
-    boostPoint: (Highcharts.Dictionary<number>|Point)
-): Point {
-    var point: Point = boostPoint as any,
-        xData = (
-            this.xData || (this.options as any).xData || this.processedXData ||
-            false
-        );
-
+Series.prototype.getPoint = function (boostPoint) {
+    var point = boostPoint, xData = (this.xData || this.options.xData || this.processedXData ||
+        false);
     if (boostPoint && !(boostPoint instanceof this.pointClass)) {
-        point = (new this.pointClass()).init( // eslint-disable-line new-cap
-            this,
-            (this.options.data as any)[boostPoint.i],
-            xData ? xData[boostPoint.i] : void 0
-        );
-
-        point.category = pick(
-            this.xAxis.categories ?
-                (this.xAxis.categories as any)[point.x as any] :
-                point.x, // @todo simplify
-            point.x
-        );
-
+        point = (new this.pointClass()).init(// eslint-disable-line new-cap
+        this, this.options.data[boostPoint.i], xData ? xData[boostPoint.i] : void 0);
+        point.category = pick(this.xAxis.categories ?
+            this.xAxis.categories[point.x] :
+            point.x, // @todo simplify
+        point.x);
         point.dist = boostPoint.dist;
         point.distX = boostPoint.distX;
         point.plotX = boostPoint.plotX;
@@ -180,118 +96,70 @@ Series.prototype.getPoint = function (
         point.index = boostPoint.i;
         point.isInside = this.isPointInside(boostPoint);
     }
-
     return point;
 };
-
 /* eslint-disable no-invalid-this */
-
 // Return a point instance from the k-d-tree
-wrap(Series.prototype, 'searchPoint', function (
-    this: Highcharts.Series,
-    proceed: Function
-): (Point|undefined) {
-    return this.getPoint(
-        proceed.apply(this, [].slice.call(arguments, 1))
-    );
+wrap(Series.prototype, 'searchPoint', function (proceed) {
+    return this.getPoint(proceed.apply(this, [].slice.call(arguments, 1)));
 });
-
 // For inverted series, we need to swap X-Y values before running base methods
-wrap(Point.prototype, 'haloPath', function (
-    this: Point,
-    proceed: Function
-): SVGPath {
-    var halo,
-        point = this,
-        series = point.series,
-        chart = series.chart,
-        plotX: number = point.plotX as any,
-        plotY: number = point.plotY as any,
-        inverted = chart.inverted;
-
+wrap(Point.prototype, 'haloPath', function (proceed) {
+    var halo, point = this, series = point.series, chart = series.chart, plotX = point.plotX, plotY = point.plotY, inverted = chart.inverted;
     if (series.isSeriesBoosting && inverted) {
         point.plotX = series.yAxis.len - plotY;
         point.plotY = series.xAxis.len - plotX;
     }
-
     halo = proceed.apply(this, Array.prototype.slice.call(arguments, 1));
-
     if (series.isSeriesBoosting && inverted) {
         point.plotX = plotX;
         point.plotY = plotY;
     }
-
     return halo;
 });
-
-wrap(Series.prototype, 'markerAttribs', function (
-    this: Highcharts.Series,
-    proceed: Function,
-    point: Point
-): Highcharts.SVGAttributes {
-    var attribs: Highcharts.SVGAttributes,
-        series = this,
-        chart = series.chart,
-        plotX: number = point.plotX as any,
-        plotY: number = point.plotY as any,
-        inverted = chart.inverted;
-
+wrap(Series.prototype, 'markerAttribs', function (proceed, point) {
+    var attribs, series = this, chart = series.chart, plotX = point.plotX, plotY = point.plotY, inverted = chart.inverted;
     if (series.isSeriesBoosting && inverted) {
         point.plotX = series.yAxis.len - plotY;
         point.plotY = series.xAxis.len - plotX;
     }
-
     attribs = proceed.apply(this, Array.prototype.slice.call(arguments, 1));
-
     if (series.isSeriesBoosting && inverted) {
         point.plotX = plotX;
         point.plotY = plotY;
     }
-
     return attribs;
 });
-
 /*
  * Extend series.destroy to also remove the fake k-d-tree points (#5137).
  * Normally this is handled by Series.destroy that calls Point.destroy,
  * but the fake search points are not registered like that.
  */
-addEvent(Series, 'destroy', function (): void {
-    var series = this,
-        chart = series.chart;
-
+addEvent(Series, 'destroy', function () {
+    var series = this, chart = series.chart;
     if (chart.markerGroup === series.markerGroup) {
-        series.markerGroup = null as any;
+        series.markerGroup = null;
     }
-
     if (chart.hoverPoints) {
-        chart.hoverPoints = chart.hoverPoints.filter(function (
-            point: Point
-        ): boolean {
+        chart.hoverPoints = chart.hoverPoints.filter(function (point) {
             return point.series === series;
         });
     }
-
     if (chart.hoverPoint && chart.hoverPoint.series === series) {
-        chart.hoverPoint = null as any;
+        chart.hoverPoint = null;
     }
 });
-
 /*
  * Do not compute extremes when min and max are set.
  * If we use this in the core, we can add the hook
  * to hasExtremes to the methods directly.
  */
-wrap(Series.prototype, 'getExtremes', function (
-    this: Highcharts.Series,
-    proceed: Function
-): Highcharts.DataExtremesObject {
+wrap(Series.prototype, 'getExtremes', function (proceed) {
     if (!this.isSeriesBoosting || (!this.hasExtremes || !this.hasExtremes())) {
         return proceed.apply(this, Array.prototype.slice.call(arguments, 1));
     }
     return {};
 });
-
 /*
  * Override a bunch of methods the same way. If the number of points is
  * below the threshold, run the original method. If not, check for a
@@ -305,37 +173,28 @@ wrap(Series.prototype, 'getExtremes', function (
     'drawTracker',
     'drawPoints',
     'render'
-].forEach(function (method: string): void {
+].forEach(function (method) {
     /**
      * @private
      */
-    function branch(
-        this: Highcharts.Series,
-        proceed: Function
-    ): void {
+    function branch(proceed) {
         var letItPass = this.options.stacking &&
             (method === 'translate' || method === 'generatePoints');
-
-        if (
-            !this.isSeriesBoosting ||
+        if (!this.isSeriesBoosting ||
             letItPass ||
             !boostEnabled(this.chart) ||
             this.type === 'heatmap' ||
             this.type === 'treemap' ||
             !boostableMap[this.type] ||
-            this.options.boostThreshold === 0
-        ) {
-
+            this.options.boostThreshold === 0) {
             proceed.call(this);
-
-        // If a canvas version of the method exists, like renderCanvas(), run
-        } else if ((this as any)[method + 'Canvas']) {
-            (this as any)[method + 'Canvas']();
+            // If a canvas version of the method exists, like renderCanvas(), run
+        }
+        else if (this[method + 'Canvas']) {
+            this[method + 'Canvas']();
         }
     }
-
     wrap(Series.prototype, method, branch);
-
     // A special case for some types - their translate method is already wrapped
     if (method === 'translate') {
         [
@@ -345,151 +204,117 @@ wrap(Series.prototype, 'getExtremes', function (
             'columnrange',
             'heatmap',
             'treemap'
-        ].forEach(function (type: string): void {
+        ].forEach(function (type) {
             if (seriesTypes[type]) {
                 wrap(seriesTypes[type].prototype, method, branch);
             }
         });
     }
 });
-
 // If the series is a heatmap or treemap, or if the series is not boosting
 // do the default behaviour. Otherwise, process if the series has no extremes.
-wrap(Series.prototype, 'processData', function (
-    this: Highcharts.Series,
-    proceed: Function
-): void {
-
-    var series = this,
-        dataToMeasure = this.options.data,
-        firstPoint: Highcharts.PointOptionsType;
-
+wrap(Series.prototype, 'processData', function (proceed) {
+    var series = this, dataToMeasure = this.options.data, firstPoint;
     /**
      * Used twice in this function, first on this.options.data, the second
      * time it runs the check again after processedXData is built.
      * @private
      * @todo Check what happens with data grouping
      */
-    function getSeriesBoosting(
-        data?: Array<Highcharts.PointOptionsType>
-    ): boolean {
-        return series.chart.isChartSeriesBoosting() || (
-            (data ? data.length : 0) >=
-            (series.options.boostThreshold || Number.MAX_VALUE)
-        );
+    function getSeriesBoosting(data) {
+        return series.chart.isChartSeriesBoosting() || ((data ? data.length : 0) >=
+            (series.options.boostThreshold || Number.MAX_VALUE));
     }
-
     if (boostEnabled(this.chart) && boostableMap[this.type]) {
-
         // If there are no extremes given in the options, we also need to
         // process the data to read the data extremes. If this is a heatmap, do
         // default behaviour.
-        if (
-            !getSeriesBoosting(dataToMeasure) || // First pass with options.data
+        if (!getSeriesBoosting(dataToMeasure) || // First pass with options.data
             this.type === 'heatmap' ||
             this.type === 'treemap' ||
             this.options.stacking || // processedYData for the stack (#7481)
             !this.hasExtremes ||
-            !this.hasExtremes(true)
-        ) {
+            !this.hasExtremes(true)) {
             proceed.apply(this, Array.prototype.slice.call(arguments, 1));
             dataToMeasure = this.processedXData;
         }
-
         // Set the isBoosting flag, second pass with processedXData to see if we
         // have zoomed.
         this.isSeriesBoosting = getSeriesBoosting(dataToMeasure);
-
         // Enter or exit boost mode
         if (this.isSeriesBoosting) {
             // Force turbo-mode:
-            firstPoint = this.getFirstValidPoint(this.options.data as any);
+            firstPoint = this.getFirstValidPoint(this.options.data);
             if (!isNumber(firstPoint) && !isArray(firstPoint)) {
                 error(12, false, this.chart);
             }
             this.enterBoost();
-        } else if (this.exitBoost) {
+        }
+        else if (this.exitBoost) {
             this.exitBoost();
         }
-
-    // The series type is not boostable
-    } else {
+        // The series type is not boostable
+    }
+    else {
         proceed.apply(this, Array.prototype.slice.call(arguments, 1));
     }
 });
-
-addEvent(Series, 'hide', function (): void {
+addEvent(Series, 'hide', function () {
     if (this.canvas && this.renderTarget) {
         if (this.ogl) {
             this.ogl.clear();
         }
         this.boostClear();
     }
-
 });
-
 /**
  * Enter boost mode and apply boost-specific properties.
  *
  * @function Highcharts.Series#enterBoost
  */
-Series.prototype.enterBoost = function (): void {
-
+Series.prototype.enterBoost = function () {
     this.alteredByBoost = [];
-
     // Save the original values, including whether it was an own property or
     // inherited from the prototype.
-    ['allowDG', 'directTouch', 'stickyTracking'].forEach(function (
-        this: Highcharts.Series,
-        prop: string
-    ): void {
-        (this.alteredByBoost as any).push({
+    ['allowDG', 'directTouch', 'stickyTracking'].forEach(function (prop) {
+        this.alteredByBoost.push({
             prop: prop,
-            val: (this as any)[prop],
+            val: this[prop],
             own: Object.hasOwnProperty.call(this, prop)
         });
     }, this);
-
     this.allowDG = false;
     this.directTouch = false;
     this.stickyTracking = true;
-
     // Prevent animation when zooming in on boosted series(#13421).
     this.finishedAnimating = true;
-
     // Hide series label if any
     if (this.labelBySeries) {
         this.labelBySeries = this.labelBySeries.destroy();
     }
 };
-
 /**
  * Exit from boost mode and restore non-boost properties.
  *
  * @function Highcharts.Series#exitBoost
  */
-Series.prototype.exitBoost = function (): void {
+Series.prototype.exitBoost = function () {
     // Reset instance properties and/or delete instance properties and go back
     // to prototype
-    (this.alteredByBoost || []).forEach(function (
-        this: Highcharts.Series,
-        setting: Highcharts.BoostAlteredObject
-    ): void {
+    (this.alteredByBoost || []).forEach(function (setting) {
         if (setting.own) {
-            (this as any)[setting.prop] = setting.val;
-        } else {
+            this[setting.prop] = setting.val;
+        }
+        else {
             // Revert to prototype
-            delete (this as any)[setting.prop];
+            delete this[setting.prop];
         }
     }, this);
-
     // Clear previous run
     if (this.boostClear) {
         this.boostClear();
     }
-
 };
-
 /**
  * @private
  * @function Highcharts.Series#hasExtremes
@@ -498,39 +323,27 @@ Series.prototype.exitBoost = function (): void {
  *
  * @return {boolean}
  */
-Series.prototype.hasExtremes = function (checkX?: boolean): boolean {
-    var options = this.options,
-        data: Array<Highcharts.PointOptionsType> = options.data as any,
-        xAxis = this.xAxis && this.xAxis.options,
-        yAxis = this.yAxis && this.yAxis.options,
-        colorAxis = this.colorAxis && this.colorAxis.options;
-
+Series.prototype.hasExtremes = function (checkX) {
+    var options = this.options, data = options.data, xAxis = this.xAxis && this.xAxis.options, yAxis = this.yAxis && this.yAxis.options, colorAxis = this.colorAxis && this.colorAxis.options;
     return data.length > (options.boostThreshold || Number.MAX_VALUE) &&
-            // Defined yAxis extremes
-            isNumber(yAxis.min) &&
-            isNumber(yAxis.max) &&
-            // Defined (and required) xAxis extremes
-            (!checkX ||
-                (isNumber(xAxis.min) && isNumber(xAxis.max))
-            ) &&
-            // Defined (e.g. heatmap) colorAxis extremes
-            (!colorAxis ||
-                (isNumber(colorAxis.min) && isNumber(colorAxis.max))
-            );
+        // Defined yAxis extremes
+        isNumber(yAxis.min) &&
+        isNumber(yAxis.max) &&
+        // Defined (and required) xAxis extremes
+        (!checkX ||
+            (isNumber(xAxis.min) && isNumber(xAxis.max))) &&
+        // Defined (e.g. heatmap) colorAxis extremes
+        (!colorAxis ||
+            (isNumber(colorAxis.min) && isNumber(colorAxis.max)));
 };
-
 /**
  * If implemented in the core, parts of this can probably be
  * shared with other similar methods in Highcharts.
  *
  * @function Highcharts.Series#destroyGraphics
  */
-Series.prototype.destroyGraphics = function (this: Highcharts.Series): void {
-    var series = this,
-        points = this.points,
-        point: Point,
-        i: number;
-
+Series.prototype.destroyGraphics = function () {
+    var series = this, points = this.points, point, i;
     if (points) {
         for (i = 0; i < points.length; i = i + 1) {
             point = points[i];
@@ -539,21 +352,17 @@ Series.prototype.destroyGraphics = function (this: Highcharts.Series): void {
             }
         }
     }
-
-    ['graph', 'area', 'tracker'].forEach(function (prop: string): void {
-        if ((series as any)[prop]) {
-            (series as any)[prop] = (series as any)[prop].destroy();
+    ['graph', 'area', 'tracker'].forEach(function (prop) {
+        if (series[prop]) {
+            series[prop] = series[prop].destroy();
         }
     });
 };
-
 // Set default options
-boostable.forEach(
-    function (type: string): void {
-        if (plotOptions[type]) {
-            (plotOptions[type] as any).boostThreshold = 5000;
-            (plotOptions[type] as any).boostData = [];
-            seriesTypes[type].prototype.fillOpacity = true;
-        }
+boostable.forEach(function (type) {
+    if (plotOptions[type]) {
+        plotOptions[type].boostThreshold = 5000;
+        plotOptions[type].boostData = [];
+        seriesTypes[type].prototype.fillOpacity = true;
     }
-);
+});
