+++ js/Extensions/Boost/BoostableMap.js
@@ -9,7 +9,7 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-import boostables from './boostables.js';
+import boostables from './Boostables.js';
 // These are the series we allow boosting for.
 var boostableMap = {};
 boostables.forEach(function (item) {
