+++ js/Extensions/Boost/NamedColors.js
