+++ js/Extensions/BoostCanvas.js
