+++ js/Series/HistogramSeries.js
