+++ js/masters/modules/boost-canvas.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../modules/boost-canvas.src.js';
+import '../../Extensions/BoostCanvas.js';
