+++ js/masters/modules/item-series.src.js
@@ -10,4 +10,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../modules/item-series.src.js';
+import '../../Series/ItemSeries.js';
