+++ ts/Extensions/Boost/BoostUtils.ts
@@ -11,16 +11,43 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
+
+import Chart from '../../Core/Chart/Chart.js';
 import H from '../../Core/Globals.js';
-var win = H.win, doc = H.doc;
+const {
+    win,
+    doc
+} = H;
+
+/**
+ * Internal types
+ * @private
+ */
+declare global {
+    namespace Highcharts {
+        /** @requires modules/boost */
+        function hasWebGLSupport(): boolean;
+        interface ChartLike {
+            boostForceChartBoost?: boolean;
+        }
+    }
+}
+
 import '../../Core/Series/Series.js';
-import boostableMap from './boostable-map.js';
-import createAndAttachRenderer from './boost-attach.js';
+import boostableMap from './BoostableMap.js';
+import createAndAttachRenderer from './BoostAttach.js';
+
 import U from '../../Core/Utilities.js';
-var pick = U.pick;
+const {
+    pick
+} = U;
+
+
 // This should be a const.
 var CHUNK_SIZE = 3000;
+
 /**
  * Tolerant max() function.
  *
@@ -33,16 +60,15 @@ var CHUNK_SIZE = 3000;
  * @return {number}
  * Max value
  */
-function patientMax() {
-    var args = [];
-    for (var _i = 0; _i < arguments.length; _i++) {
-        args[_i] = arguments[_i];
-    }
+function patientMax(...args: Array<Array<unknown>>): number {
     var r = -Number.MAX_VALUE;
-    args.forEach(function (t) {
-        if (typeof t !== 'undefined' &&
+
+    args.forEach(function (t: Array<unknown>): (boolean|undefined) {
+        if (
+            typeof t !== 'undefined' &&
             t !== null &&
-            typeof t.length !== 'undefined') {
+            typeof t.length !== 'undefined'
+        ) {
             // r = r < t.length ? t.length : r;
             if (t.length > 0) {
                 r = t.length;
@@ -50,8 +76,10 @@ function patientMax() {
             }
         }
     });
+
     return r;
 }
+
 /**
  * Return true if ths boost.enabled option is true
  *
@@ -64,12 +92,18 @@ function patientMax() {
  * @return {boolean}
  * True, if boost is enabled.
  */
-function boostEnabled(chart) {
-    return pick((chart &&
-        chart.options &&
-        chart.options.boost &&
-        chart.options.boost.enabled), true);
+function boostEnabled(chart: Chart): boolean {
+    return pick(
+        (
+            chart &&
+            chart.options &&
+            chart.options.boost &&
+            chart.options.boost.enabled
+        ),
+        true
+    );
 }
+
 /**
  * Returns true if we should force boosting the chart
  * @private
@@ -81,16 +115,26 @@ function boostEnabled(chart) {
  * @return {boolean}
  * True, if boosting should be forced.
  */
-function shouldForceChartSeriesBoosting(chart) {
+function shouldForceChartSeriesBoosting(chart: Chart): boolean {
     // If there are more than five series currently boosting,
     // we should boost the whole chart to avoid running out of webgl contexts.
-    var sboostCount = 0, canBoostCount = 0, allowBoostForce = pick(chart.options.boost && chart.options.boost.allowForce, true), series;
+    var sboostCount = 0,
+        canBoostCount = 0,
+        allowBoostForce = pick(
+            chart.options.boost && chart.options.boost.allowForce,
+            true
+        ),
+        series;
+
     if (typeof chart.boostForceChartBoost !== 'undefined') {
         return chart.boostForceChartBoost;
     }
+
     if (chart.series.length > 1) {
         for (var i = 0; i < chart.series.length; i++) {
+
             series = chart.series[i];
+
             // Don't count series with boostThreshold set to 0
             // See #8950
             // Also don't count if the series is hidden.
@@ -99,28 +143,42 @@ function shouldForceChartSeriesBoosting(chart) {
                 series.visible === false) {
                 continue;
             }
+
             // Don't count heatmap series as they are handled differently.
             // In the future we should make the heatmap/treemap path compatible
             // with forcing. See #9636.
             if (series.type === 'heatmap') {
                 continue;
             }
+
             if (boostableMap[series.type]) {
                 ++canBoostCount;
             }
-            if (patientMax(series.processedXData, series.options.data, 
-            // series.xData,
-            series.points) >= (series.options.boostThreshold || Number.MAX_VALUE)) {
+
+            if (patientMax(
+                series.processedXData,
+                series.options.data as any,
+                // series.xData,
+                series.points
+            ) >= (series.options.boostThreshold || Number.MAX_VALUE)) {
                 ++sboostCount;
             }
         }
     }
-    chart.boostForceChartBoost = allowBoostForce && ((canBoostCount === chart.series.length &&
-        sboostCount > 0) ||
-        sboostCount > 5);
+
+    chart.boostForceChartBoost = allowBoostForce && (
+        (
+            canBoostCount === chart.series.length &&
+            sboostCount > 0
+        ) ||
+        sboostCount > 5
+    );
+
     return chart.boostForceChartBoost;
 }
+
 /* eslint-disable valid-jsdoc */
+
 /**
  * Performs the actual render if the renderer is
  * attached to the series.
@@ -128,25 +186,36 @@ function shouldForceChartSeriesBoosting(chart) {
  * @param renderer {OGLRenderer} - the renderer
  * @param series {Highcharts.Series} - the series
  */
-function renderIfNotSeriesBoosting(renderer, series, chart) {
+function renderIfNotSeriesBoosting(
+    renderer: Highcharts.BoostGLRenderer,
+    series: Highcharts.Series,
+    chart?: Chart
+): void {
     if (renderer &&
         series.renderTarget &&
         series.canvas &&
-        !(chart || series.chart).isChartSeriesBoosting()) {
+        !(chart || series.chart).isChartSeriesBoosting()
+    ) {
         renderer.render(chart || series.chart);
     }
 }
+
 /**
  * @private
  */
-function allocateIfNotSeriesBoosting(renderer, series) {
+function allocateIfNotSeriesBoosting(
+    renderer: Highcharts.BoostGLRenderer,
+    series: Highcharts.Series
+): void {
     if (renderer &&
         series.renderTarget &&
         series.canvas &&
-        !series.chart.isChartSeriesBoosting()) {
+        !series.chart.isChartSeriesBoosting()
+    ) {
         renderer.allocateBufferForSingleSeries(series);
     }
 }
+
 /**
  * An "async" foreach loop. Uses a setTimeout to keep the loop from blocking the
  * UI thread.
@@ -160,36 +229,47 @@ function allocateIfNotSeriesBoosting(renderer, series) {
  * @param i {Number} - the current index
  * @param noTimeout {Boolean} - set to true to skip timeouts
  */
-function eachAsync(arr, fn, finalFunc, chunkSize, i, noTimeout) {
+function eachAsync(
+    arr: Array<unknown>,
+    fn: Function,
+    finalFunc: Function,
+    chunkSize?: number,
+    i?: number,
+    noTimeout?: boolean
+): void {
     i = i || 0;
     chunkSize = chunkSize || CHUNK_SIZE;
-    var threshold = i + chunkSize, proceed = true;
+
+    var threshold = i + chunkSize,
+        proceed = true;
+
     while (proceed && i < threshold && i < arr.length) {
         proceed = fn(arr[i], i);
         ++i;
     }
+
     if (proceed) {
         if (i < arr.length) {
+
             if (noTimeout) {
                 eachAsync(arr, fn, finalFunc, chunkSize, i, noTimeout);
-            }
-            else if (win.requestAnimationFrame) {
+            } else if (win.requestAnimationFrame) {
                 // If available, do requestAnimationFrame - shaves off a few ms
-                win.requestAnimationFrame(function () {
+                win.requestAnimationFrame(function (): void {
+                    eachAsync(arr, fn, finalFunc, chunkSize, i);
+                });
+            } else {
+                setTimeout(function (): void {
                     eachAsync(arr, fn, finalFunc, chunkSize, i);
                 });
             }
-            else {
-                setTimeout(function () {
-                    eachAsync(arr, fn, finalFunc, chunkSize, i);
-                });
-            }
-        }
-        else if (finalFunc) {
+
+        } else if (finalFunc) {
             finalFunc();
         }
     }
 }
+
 /**
  * Returns true if the current browser supports webgl
  *
@@ -198,25 +278,32 @@ function eachAsync(arr, fn, finalFunc, chunkSize, i, noTimeout) {
  *
  * @return {boolean}
  */
-function hasWebGLSupport() {
-    var i = 0, canvas, contexts = ['webgl', 'experimental-webgl', 'moz-webgl', 'webkit-3d'], context = false;
+function hasWebGLSupport(): boolean {
+    var i = 0,
+        canvas,
+        contexts = ['webgl', 'experimental-webgl', 'moz-webgl', 'webkit-3d'],
+        context: (false|WebGLRenderingContext|null) = false;
+
     if (typeof win.WebGLRenderingContext !== 'undefined') {
         canvas = doc.createElement('canvas');
+
         for (; i < contexts.length; i++) {
             try {
-                context = canvas.getContext(contexts[i]);
+                context = canvas.getContext(contexts[i]) as any;
                 if (typeof context !== 'undefined' && context !== null) {
                     return true;
                 }
-            }
-            catch (e) {
+            } catch (e) {
                 // silent error
             }
         }
     }
+
     return false;
 }
+
 /* eslint-disable no-invalid-this */
+
 /**
  * Used for treemap|heatmap.drawPoints
  *
@@ -227,26 +314,35 @@ function hasWebGLSupport() {
  *
  * @return {*}
  */
-function pointDrawHandler(proceed) {
-    var enabled = true, renderer;
+function pointDrawHandler(this: Highcharts.Series, proceed: Function): void {
+    var enabled = true,
+        renderer: Highcharts.BoostGLRenderer;
+
     if (this.chart.options && this.chart.options.boost) {
         enabled = typeof this.chart.options.boost.enabled === 'undefined' ?
             true :
             this.chart.options.boost.enabled;
     }
+
     if (!enabled || !this.isSeriesBoosting) {
         return proceed.call(this);
     }
+
     this.chart.isBoosting = true;
+
     // Make sure we have a valid OGL context
     renderer = createAndAttachRenderer(this.chart, this);
+
     if (renderer) {
         allocateIfNotSeriesBoosting(renderer, this);
         renderer.pushSeries(this);
     }
+
     renderIfNotSeriesBoosting(renderer, this);
 }
+
 /* eslint-enable no-invalid-this, valid-jsdoc */
+
 var funs = {
     patientMax: patientMax,
     boostEnabled: boostEnabled,
@@ -257,6 +353,8 @@ var funs = {
     hasWebGLSupport: hasWebGLSupport,
     pointDrawHandler: pointDrawHandler
 };
+
 // This needs to be fixed.
 H.hasWebGLSupport = hasWebGLSupport;
+
 export default funs;
