+++ ts/Extensions/Boost/NamedColors.ts
