+++ ts/Extensions/Boost/WGLRenderer.ts
@@ -108,8 +108,8 @@ declare global {
     }
 }
 
-import GLShader from './wgl-shader.js';
-import GLVertexBuffer from './wgl-vbuffer.js';
+import GLShader from './WGLShader.js';
+import GLVertexBuffer from './WGLVBuffer.js';
 import Color from '../../Core/Color.js';
 const color = Color.parse;
 import U from '../../Core/Utilities.js';
