+++ ts/Series/BellcurveSeries.ts
@@ -12,8 +12,6 @@
 
 'use strict';
 
-import H from '../Core/Globals.js';
-
 /**
  * Internal types
  * @private
